tag.c: use type_from_string_gently() when parsing tags
Change a series of strcmp() to instead use type_from_string_gently()
to get the integer type early, and then use that for comparison. By
making use of type_from_string_gently() we can avoid the memcpy to the
local "char type[20]". There's no need to hardcode the maximum length
of tags here.

In converting this code from the series of strcmp() we also need to
have the "unknown tag type" handling be contingent of the return value
of type_from_string_gently(). That function guarantees that it'll
return either "< 0", or a valid type that if given to type_name() will
yield a non-NULL return value, so we can change the "else" branch to
call BUG(...).

This changes the behavior errors we'd get if the tag length exceeded
the previous "char type[20]" variable. We'd previously silently
"return -1" in those cases, now we'll additionally emit an an "unknown
tag type" error message.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>