Makefile: split up the deceleration of PERL_DEFINES
Split the declaration of PERL_DEFINES across multiple line, making
this easier to read.

In 07d90eadb50 (Makefile: add Perl runtime prefix support, 2018-04-10)
when PERL_DEFINES was added only the RUNTIME_PREFIX was put on its own
line the rest weren't formatted like that for consistency. Let's do
that to make this consistent with most of the rest of this Makefile.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>