Makefile: don't re-define PERL_DEFINES
Since 07d90eadb50 (Makefile: add Perl runtime prefix support,
2018-04-10) we have been declaring PERL_DEFINES right after assigning
to it, with the effect that the first PERL_DEFINES was ignored.

This didn't matter in practice since the first line had all the same
variables as the second, so we'd correctly re-generate everything. It
just made for confusing reading. Let's remove that first assignment.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>