C99 support: hard-depend on C99 variadic macros
Remove the "else" branches of the HAVE_VARIADIC_MACROS macro, which
has been unconditionally set since 765dc168882 (git-compat-util:
always enable variadic macros, 2021-01-28).

That commit went out with v2.31.0, so we've got a few releases already
where we've been explicitly breaking things for compilers that don't
support this C99 feature.

In addition to that we've been emitting extremely verbose warnings
since at least ee4512ed481 (trace2: create new combined trace
facility, 2019-02-22) if HAVE_VARIADIC_MACROS wasn't true. There is no
such thing as a "region_enter_printf" or "region_leave_printf" format,
so at least under GCC and Clang everything that includes
trace.h (almost every file) emits a couple of warnings about that.

There's a large benefit to being able to rely on variadic macros, the
code surrounding usage.c is hard to maintain if we need to write two
implementations of everything, and by relying on "__FILE__" and
"__LINE__" along with "__VA_ARGS__" we can in the future make error(),
die() etc. log where they were called from, without the verbosity of a
dual-implementation.

Let's also update our CodingGuidelines to note that we depend on
this. The added bullet-point starts with lower-case for consistency
with other bullet-points in that section.

The diff in "trace.h" is relatively hard to read, since we need to
retain the existing API docs, which were comments on the code used if
HAVE_VARIADIC_MACROS was not defined.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>