test-lib tests: filter out verbose "git init" output
Filter out the output emitted by "git init" when not under the
"--quiet" option, as is done when test-lib.sh itself invokes it, or
when "git init" or "test_create_repo()" are called.

The output contains the absolute path to the initialized repository,
so it can't be test_cmp'd without munging it. Let's instead ignore it
for now. Ignoring it means that any test using "t/lib-subtest.sh" with
"--verbose" won't need to manually ignore it.

As can be seen here this part of the output was already being removed
in a test added in 517cd55fd51 (test-lib: self-test that --verbose
works, 2013-06-23). That code won't be needed anymore, as we're now
doing the equivalent of it in "t/lib-subtest.sh" instead.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>