test-lib output: emit pretty "git init" output on setup
Replace the "Initialized.* Git repository" output which includes the
absolute path to the repository with TAP-like trace output.

This allows for reverting the grepping out of the "git init" output
from lib-subtest.sh, we can now "test_cmp" this output. So the output
isn't new as the test changes might suggest, just changed.

This also had an interaction with the --verbose-only=*, it was only
prepared to handle verbose output emitted after we'd invoked the first
test via test_expect_{success,failure}. So let's call
"maybe_setup_verbose" earlier.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>