midx: call display_progress() at the start of loops
Change code in midx.c to consistently call display_progress() at the
start of loops. In some of these we "continue" before counting the
work we're doing, but mostly this is just for ease of reading the
code. We don't need to worry about a progress bar not having shown at
all because we're still hanging on the first item.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>