config.mak: avoid quoting issues between t/Makefile and t/perf/run
In the t/perf/run script we source GIT-BUILD-OPTIONS and then do:

    ./$t $GIT_TEST_OPTS

But in the t/Makefile we do:

	@echo "*** prove ***"; $(PROVE) --exec '$(TEST_SHELL_PATH_SQ)' $(GIT_PROVE_OPTS) $(T) :: $(GIT_TEST_OPTS)

The result of the two is that having "" here is passed as-is in the
perf script, but not in the t/Makefile. I want to just not have the
quotes here, so let's avoid them.